requirements.txt
flask==3.1.1
numpy==2.1.3
pillow==11.1.0
scikit-image==0.25.2
tensorflow==2.19.0
matplotlib==3.10.0